Moved test class to test files and tested a bit more within DataSupplierImpl
